IntuitIQ-be/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
pymongo==4.11.3
pytz==2025.2
python-dotenv==1.0.1
Pillow==11.0.0
google-genai==1.19.0
mistralai==1.5.0